requirements.txt
fastapi
uvicorn
python-decouple
SQLAlchemy
PyMySQL
alembic
pyjwt
passlib
email-validator
python-jose
python-multipart
sendgrid
